requirements.txt
flask
streamlit
python-dotenv
Flask-SQLAlchemy
Werkzeug
Flask-JWT-Extended